Replace global sidebar nav with page navigation
Since the global navigation is available from the masthead, there's no need to duplicate it in the sidebar. This allows for inclusion of in-page navigation between different page sections.

Due to how Jekyll stores everything in files, each page needs to have its own set of in-page anchors declared in a separate file, under `/docs/_includes/nav/`.

Added first batch of navigation includes.

Refs #14