[Docs] Hide `masthead` & `signup` pages
Masthead component is currently too unstable to include it within the WFP UI documentation, as it's pending review and rebuild. It was temporarily hidden, until the issues are resolved.

It was decided that the sign-up widget will be removed, but the documentation for it needs to end up somewhere, thus it was hidden, until we can find it a better place.

Refs #14